+++ CHANGELOG.md
@@ -8,9 +8,9 @@ The *alited* has acquired the rich and comprehensive set of features, some are u
 # LAST CHANGES
 
 
-Version `v1.8.2a1 (29 Mar'24)`
+Version `v1.8.2a2 (30 Mar'24)`
 
-  - NEW   : Edit/Formats: mode=6: pluginables; "Info" as sample
+  - NEW   : Edit/Formats: mode=6 pluginables: "Info" & "Calendar" as samples
   - CHANGE: infobar: colors of results and found unit names (branches)
   - CHANGE: Edit/Formats: mode=5: %w (text path) and %v (selection) wildcards
   - CHANGE: packages: e_menu/apave 4.4.1